Add decoupled unit of work pattern to allow for multiple DbContexts to be used in the same application via IUnitOfWork Rename EfUnitOfWork to AppUnitOfWork Fix issue with long to string converter when null or empty string is passed in => now defaults to 0 instead of throwing an exception Adding support for ProblemHttpResult in the TypedResultsBuilder
