+++ src/Endatix.Infrastructure/Identity/Authentication/JwtTokenService.cs
@@ -51,8 +51,8 @@ internal sealed class JwtTokenService : IUserTokenService
         var credentials = new SigningCredentials(securityKey, SecurityAlgorithms.HmacSha256);
         var subject = new ClaimsIdentity(claims: [
                 new Claim(JwtRegisteredClaimNames.Sub, forUser.Id.ToString()),
+                new Claim(ClaimTypes.NameIdentifier, forUser.Id.ToString()),
                 new Claim(JwtRegisteredClaimNames.Email, forUser.Email.ToString()),
-                new Claim(ClaimTypes.NameIdentifier, forUser.UserName.ToString()),
                 new Claim(ClaimTypes.Role, RoleNames.ADMIN),
                 new Claim(ClaimNames.Permission, Allow.AllowAll),
                 new Claim(ClaimNames.TenantId, forUser.TenantId.ToString())
